rvme/requirements/local.txt
-r common.txt

# local utils
Werkzeug
django-debug-toolbar
django-extensions